feat(t11): inspection + capital deployment systems
- inspectionEngine + InspectionPanel (dialog with focus trap, PostureSelector radiogroup; STRATEGIC_ALTERNATIVE gated by precondition flag)
- capitalResolver + CapitalPowerCard (6-verb radiogroup + explicit Commit; never auto-commit)
- q1 content: inspection scene + capital power cards
- inspectionSlice, capitalSlice, flagsSlice (game-wide flags)
- 214/214 tests pass, typecheck + lint clean

+++ game/src/schemas/choiceNode.schema.ts
@@ -19,7 +19,7 @@ export const ChoiceNodeSchema = z.object({
   label: z.string().min(1),
   keybind: z.enum(['1', '2', '3', '4']),
   /** Per-meter integer changes; absent keys mean no change. */
-  meterDeltas: z.partialRecord(MeterKeySchema, z.number()),
+  meterDeltas: z.partialRecord(MeterKeySchema, z.number().int()),
   /** ConsequenceHook ids scheduled when this choice is taken. */
   scheduledConsequenceIds: z.array(ConsequenceIdSchema),
 })
